+++ examples/CreditCardAuthorization/Controller.php
@@ -29,6 +29,7 @@ require_once __DIR__ . '/Constants.php';
 /** Require the composer autoloader file */
 require_once __DIR__ . '/../../../../autoload.php';
 
+use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
 use heidelpayPHP\Resources\Customer;
@@ -52,10 +53,12 @@ $paymentTypeId   = $_POST['resourceId'];
 try {
     //#######  2. Create a heidelpay object using your private key #####################################################
     $heidelpay = new Heidelpay('s-priv-2a102ZMq3gV4I3zJ888J7RR6u75oqK3n');
+    $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
     //#######  3. Create an authorization (aka reservation) ############################################################
-    $customer      = new Customer('Linda', 'Heideich');
-    $authorization = $heidelpay->authorize(12.99, 'EUR', $paymentTypeId, CONTROLLER_URL, $customer);
+    $customer            = new Customer('Linda', 'Heideich');
+    $authorization       = $heidelpay->authorize(12.99, 'EUR', $paymentTypeId, CONTROLLER_URL, $customer);
+    $_SESSION['ShortId'] = $authorization->getShortId();
     if ($authorization->getPayment()->isPending()) {
         redirect(SUCCESS_URL);
     }