+++ src/Resources/TransactionTypes/Charge.php
@@ -25,8 +25,6 @@
 namespace heidelpayPHP\Resources\TransactionTypes;
 
 use heidelpayPHP\Exceptions\HeidelpayApiException;
-use heidelpayPHP\Resources\Payment;
-use heidelpayPHP\Resources\PaymentTypes\BasePaymentType;
 use heidelpayPHP\Traits\HasCancellations;
 
 class Charge extends AbstractTransactionType
@@ -127,26 +125,6 @@ class Charge extends AbstractTransactionType
     protected function getResourcePath(): string
     {
         return 'charges';
-    }
-
-    /**
-     * {@inheritDoc}
-     *
-     * @throws \RuntimeException
-     */
-    public function getLinkedResources(): array
-    {
-        /** @var Payment $payment */
-        $payment = $this->getPayment();
-        $paymentType = $payment ? $payment->getPaymentType() : null;
-        if (!$paymentType instanceof BasePaymentType) {
-            throw new \RuntimeException('Payment type is missing!');
-        }
-
-        return [
-            'customer'=> $payment->getCustomer(),
-            'type' => $paymentType
-        ];
     }
 
     //</editor-fold>